Remove custom StackFrame type
This removes the direct dependency between all implementations of the
stackTracer interface and newrelic. This change should remain compatible
with older go versions
